fastapi_app_be/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic>=2.6.0
pydantic-core>=2.16.0
pydantic-settings==2.1.0
google-generativeai==0.3.0
pypdf==4.0.1
python-docx==0.8.11
python-pptx==0.6.23
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
requests==2.31.0
httpx==0.25.2
PyPDF2
